Correction : Prise en compte des versions ancienne de webdelib

+++ trunk/cake_webdelib/Config/Schema/patchs/4.1_to_4.2.sql
@@ -107,7 +107,7 @@ ALTER TABLE deliberations ALTER COLUMN tdt_id TYPE VARCHAR;
 ALTER TABLE deliberations ADD COLUMN parapheur_bordereau BYTEA DEFAULT NULL;
 
 -- Table Nomenclatures
-DROP TABLE nomenclatures;
+IF EXISTS DROP TABLE nomenclatures;
 CREATE TABLE nomenclatures (
   id        VARCHAR PRIMARY KEY         NOT NULL,
   parent_id VARCHAR                     NOT NULL DEFAULT 0,
